+++ tests/Client1.Test.php
@@ -3,9 +3,7 @@ use PHPUnit\Framework\TestCase;
 use McFish\LibMQTT\Client;
 
 class Client1 extends TestCase {
-
 	public function testUnencryptedClientCreation() {
-
 		// Try to establish connection to server
 		$client = new Client("localhost",1883,"phpUnitClient");
 		$client->setAuthDetails("testuser", "userpass");
@@ -31,12 +29,11 @@ class Client1 extends TestCase {
 
 	public function testCryptedClientCreation() {
 
-		// Try to establish connection to server
 		$client = new Client("test.mosquitto.org",8883,"phpUnitClient");
-		$client->setCryptoProtocol("tls");
+		$client->setTransportProtocol("tls");
 		$client->setCAFile("tests/mosquitto.org.crt");
 
-		//
+		// Try to establish connection to server
 		$result = $client->connect();
 		$this->assertTrue($result);
 
@@ -47,11 +44,11 @@ class Client1 extends TestCase {
 		] );
 		$this->assertTrue($result);
 
-		//
+		// Try to publish message
 		$result = $client->publish("libmqtt/test", "testi", 0);
  		$this->assertTrue($result);
 
-		//
+		// Close connection
 		$client->close();
 
 	}